requirements.txt
Django==3.0.5
django-filter==2.2.0
djangorestframework==3.11.0
docutils==0.16
gunicorn==20.0.4
Jinja2==2.10.1
numpy==1.18.3
pandas==1.0.3
Pillow==7.0.0
psycopg2==2.8.5
